FEAT/#21: Data 모듈 추가, Service 추가
