+++ data/src/main/java/com/sopt/now/compose/data/service/HomeService.kt
@@ -1,0 +1,11 @@
+package com.sopt.now.compose.data.service
+
+
+import com.sopt.now.compose.data.dto.response.ResponseUserDto
+import retrofit2.Response
+import retrofit2.http.GET
+
+interface HomeService {
+    @GET("member/info")
+    suspend fun getMemberInfo(): Response<ResponseUserDto>
+}
